Move UserLogin to BaseTest class
