+++ ui-tests/src/test/java/com/practiseofsoftwaretesting/BaseTest.java
@@ -1,12 +1,22 @@
 package com.practiseofsoftwaretesting;
 
 import com.codeborne.selenide.Configuration;
+import com.practicesoftwaretesting.pages.AccountPage;
+import com.practicesoftwaretesting.pages.HomePage;
+import com.practicesoftwaretesting.pages.LoginPage;
+import com.practicesoftwaretesting.user.UserSteps;
 import org.junit.jupiter.api.AfterEach;
 
 import static com.codeborne.selenide.Selenide.closeWebDriver;
 import com.practicesoftwaretesting.utils.ConfigReader;
+import static com.practicesoftwaretesting.user.UserSteps.DEFAULT_PASSWORD;
 
 public abstract class BaseTest {
+
+    UserSteps userSteps = new UserSteps();
+    LoginPage loginPage = new LoginPage();
+    AccountPage accountPage = new AccountPage();
+    HomePage homePage = new HomePage();
 
     static ConfigReader configReader = new ConfigReader();
 
@@ -18,6 +28,24 @@ public abstract class BaseTest {
         Configuration.fastSetValue = Boolean.parseBoolean(configReader.getProperty("fast.set.value"));
         Configuration.headless = Boolean.parseBoolean(configReader.getProperty("headless"));
         Configuration.proxyEnabled = Boolean.parseBoolean(configReader.getProperty("proxy.enabled"));
+    }
+
+    public void registerAndLoginAsNewUser() {
+        var email = userSteps.getUserEmail();
+        userSteps.registerUser(email, DEFAULT_PASSWORD);
+
+        loginPage.open()
+                .isLoaded()
+                .login(email, DEFAULT_PASSWORD);
+        accountPage.isLoaded();
+        homePage.open();
+    }
+
+    public void loginAsAdmin() {
+        loginPage.open()
+                .isLoaded()
+                .login(UserSteps.ADMIN_EMAIL, UserSteps.ADMIN_PASSWORD);
+        accountPage.isLoaded();
     }
 
     @AfterEach
